+++ jobhelper/jobhelper.lua
@@ -13,7 +13,7 @@
 
 addon.name      = 'jobhelper';
 addon.author    = 'watahero';
-addon.version   = '2.1.4';
+addon.version   = '2.1.5';
 addon.desc      = 'RUN/PUP/NIN helpers in one compact bar for CatsEyeXI. Fork of runehelper, puphelper and ninhelper by GetAwayCoxn.';
 addon.link      = 'https://github.com/watahero/jobhelper';
 
@@ -126,6 +126,7 @@ local function BuildLightContext()
     local player = memory:GetPlayer();
 
     ctx.player = player;
+    ctx.armed = state.enabled;
     ctx.now = os.time();
     ctx.mainJob = player:GetMainJob();
     ctx.subJob = player:GetSubJob();